# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 临床中心台 SiteDesk API
  version: 0.2.0
  description: |-
    **本文件是产物。** 契约的唯一定义源是 `packages/contracts/src` 下的 zod schema。

    分层：L1 资源层（CRUD 与列表）· L2 命令层（领域动作，返回 sideEffects）· L3 投影层（只读聚合，永不返回明细）。

    带 `x-gated-by` 的字段受列维度权限管辖：**无权限时该字段从响应中消失，不是返回 null**。

    行范围之外的资源一律返回 404 而非 403 —— 403 等于确认「它存在，只是你不能碰」，这个确认本身就是泄漏。
servers:
  - url: https://api.sitedesk.example
    description: 生产
...
paths:
  /v1/accounts:
    get:
      operationId: listAccounts
      tags:
        - identity
      summary: 账号列表
      description: 外部角色只看得到自己 —— 由行级安全强制，不靠调用方传参过滤。
      x-layer: L1
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - name: limit
          in: query
          required: false
          schema:
            default: 50
            type: integer
            minimum: 1
            maximum: 200
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - active
              - disabled
        - name: roleCode
          in: query
          required: false
          schema:
            type: string
        - name: q
          in: query
          required: false
          description: 按姓名或登录名模糊匹配
          schema:
            type: string
            maxLength: 64
      responses:
        '200':
          description: 账号列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAccountsResponse'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
    post:
      operationId: createAccount
      tags:
        - identity
      summary: 新增账号
      description: |-
        不设密码：内部走 OIDC（企业微信 / 飞书），外部走一次性魔法链接。
        row_rule=hospital 的角色必须同时给出 orgRef，否则账号能登录却一行数据都看不到。
      x-layer: L1
      x-required-action: manage
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: 新增账号
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
  /v1/accounts/{id}:disable:
    post:
      operationId: disableAccount
      tags:
        - identity
      summary: 停用账号
      description: |-
        停用不删除 —— 审计轨迹必须能追溯到人。
        **仍带着中心的人不能直接停用**：必须先发起交接，否则那些中心会失去负责人而无人察觉。
      x-layer: L2
      x-required-action: manage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: >-
              ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        - name: Idempotency-Key
          in: header
          required: true
          description: 幂等键（uuid）。24 小时内重放同一键返回首次结果。
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisableAccountRequest'
      responses:
        '200':
          description: 停用账号
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisableAccountResponse'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '409':
          description: 版本冲突：资源已被他人修改
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/conflict-version
                title: 版本冲突：资源已被他人修改
                status: 409
                code: conflict-version
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
  /v1/audit-entries:
    get:
      operationId: listAuditEntries
      tags:
        - identity
      summary: 审计轨迹
      description: 只追加、不可改删。外部方只看得到本院中心相关的条目。
      x-layer: L1
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - name: limit
          in: query
          required: false
          schema:
            default: 50
            type: integer
            minimum: 1
            maximum: 200
        - name: studySiteId
          in: query
          required: false
          schema:
            type: string
            format: uuid
            pattern: >-
              ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        - name: actorLogin
          in: query
          required: false
          schema:
            type: string
        - name: targetType
          in: query
          required: false
          schema:
            type: string
        - name: targetId
          in: query
          required: false
          schema:
            type: string
        - name: sensitiveOnly
          in: query
          required: false
          description: 只看权限类变更
          schema:
            type: boolean
        - name: since
          in: query
          required: false
          schema:
            type: string
            format: date-time
            pattern: >-
              ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
      responses:
        '200':
          description: 审计轨迹
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAuditEntriesResponse'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
...
  /v1/me:
    get:
      operationId: getMe
      tags:
        - identity
      summary: 当前主体与权限
      description: |-
        前端启动时的第一个调用。导航、字段遮罩、按钮可用性全部由它驱动。
        **前端不得自行推断权限** —— 它只负责按这里给的结果收敛 UI；服务端会独立再强制一次。
      x-layer: L1
      parameters: []
      responses:
        '200':
          description: 当前主体与权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Principal'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
...
  /v1/roles:
    get:
      operationId: listRoles
      tags:
        - identity
      summary: 角色与三维权限
      x-layer: L1
      parameters: []
      responses:
        '200':
          description: 角色与三维权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRolesResponse'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
  /v1/roles/{id}:
    patch:
      operationId: updateRolePermissions
      tags:
        - identity
      summary: 调整角色权限
      description: |-
        行 / 列 / 动作三个维度可分别调整，改完对该角色的所有账号**立即生效**。
        每次调整都写审计，且 isSensitive=true —— 「谁给谁开了什么」是核查必查项。
      x-layer: L1
      x-required-action: manage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: >-
              ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRolePermissionsRequest'
      responses:
        '200':
          description: 调整角色权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '409':
          description: 版本冲突：资源已被他人修改
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/conflict-version
                title: 版本冲突：资源已被他人修改
                status: 409
                code: conflict-version
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
...
components:
  schemas:
    FieldIssue:
      type: object
      properties:
        path:
          type: string
          description: JSON Pointer，如 /hours
        message:
          type: string
      required:
        - path
        - message
      additionalProperties: false
    GateUnmet:
      type: object
      properties:
        code:
          type: string
          description: 如 open-queries / ip-imbalance
        message:
          type: string
          description: 面向人的说明，如「3 条数据质疑未关闭」
        module:
          description: 处理入口所属模块，供前端给出「去处理」跳转
          type: string
      required:
        - code
        - message
      additionalProperties: false
    Problem:
      type: object
      properties:
        type:
          type: string
          description: https://sitedesk.dev/problems/{code}
        title:
          type: string
        status:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        code:
          type: string
          enum:
            - validation-failed
            - unauthenticated
            - forbidden-action
            - not-found
            - conflict-version
            - idempotency-key-reused
            - invariant-violated
            - gate-not-satisfied
            - rate-limited
            - internal
        detail:
          type: string
        instance:
          description: 出错的请求路径
          type: string
        traceId:
          type: string
        issues:
          type: array
          items:
            $ref: '#/components/schemas/FieldIssue'
        unmet:
          type: array
          items:
            $ref: '#/components/schemas/GateUnmet'
        invariant:
          type: string
      required:
        - type
        - title
        - status
        - code
      additionalProperties: false
      description: RFC 9457 Problem Details。所有非 2xx 响应都是这个形状。
...
    RowRule:
      type: string
      enum:
        - all
        - team
        - assigned
        - hospital
        - pi
        - none
      description: 行范围规则。all=全部｜team=本组承接的项目｜assigned=被指派的中心｜hospital=本院承接的项目｜pi=本人担任研究者的中心｜none=无。**由身份推导，绝不由用户选择。**
...
    AccountStatus:
      type: string
      enum:
        - active
        - disabled
...
    Account:
      type: object
      properties:
        id:
          type: string
          format: uuid
          pattern: >-
            ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        login:
          type: string
        displayName:
          type: string
        role:
          $ref: '#/components/schemas/RoleRef'
        team:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/TeamRef'
        isExternal:
          type: boolean
        orgRef:
          nullable: true
          description: 外部方所属机构。row_rule=hospital 的账号必填，否则行范围为空。
          type: string
        status:
          $ref: '#/components/schemas/AccountStatus'
        joinedOn:
          nullable: true
          type: string
          format: date
          pattern: >-
            ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
        disabledAt:
          nullable: true
          type: string
          format: date-time
          pattern: >-
            ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
        disabledReason:
          nullable: true
          type: string
        lastLoginAt:
          nullable: true
          type: string
          format: date-time
          pattern: >-
            ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
      required:
        - id
        - login
        - displayName
        - role
        - team
        - isExternal
        - orgRef
        - status
        - joinedOn
        - disabledAt
        - disabledReason
        - lastLoginAt
      additionalProperties: false
    Principal:
      type: object
      properties:
        account:
          $ref: '#/components/schemas/Account'
        scopeLabel:
          type: string
        visibleSiteIds:
          type: array
          items:
            type: string
            format: uuid
            pattern: >-
              ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          description: 当前可见的中心 id 全集。前端用于本地过滤；**服务端仍会独立强制一次**。
        permissions:
          type: object
          properties:
            rowRule:
              $ref: '#/components/schemas/RowRule'
            fields:
              type: array
              items:
                type: string
                enum:
                  - cost
                  - margin
                  - price
                  - subject
                  - staff
            actions:
              type: array
              items:
                $ref: '#/components/schemas/ActionKey'
            modules:
              type: array
              items:
                type: string
          required:
            - rowRule
            - fields
            - actions
            - modules
          additionalProperties: false
      required:
        - account
        - scopeLabel
        - visibleSiteIds
        - permissions
      additionalProperties: false
...
    ListAccountsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        nextCursor:
          nullable: true
          description: null 表示已到末页
          type: string
          minLength: 1
          maxLength: 512
      required:
        - items
        - nextCursor
      additionalProperties: false
    CreateAccountRequest:
      type: object
      properties:
        login:
          type: string
          pattern: ^[a-z][a-z0-9_]{2,31}$
        displayName:
          type: string
          minLength: 1
          maxLength: 64
        roleId:
          type: string
          format: uuid
          pattern: >-
            ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        teamId:
          nullable: true
          type: string
          format: uuid
          pattern: >-
            ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        orgRef:
          nullable: true
          type: string
          maxLength: 128
      required:
        - login
        - displayName
        - roleId
      additionalProperties: false
    DisableAccountRequest:
      type: object
      properties:
        reason:
          type: string
          minLength: 4
          maxLength: 500
          description: 变更原因。会写入审计轨迹；核查时真正被问的就是这一栏。
      required:
        - reason
      additionalProperties: false
    DisableAccountResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Account'
        sideEffects:
          type: array
          items:
            $ref: '#/components/schemas/SideEffect'
          description: 本次命令连带触发的事件。为空数组表示没有连带影响。
      required:
        - data
        - sideEffects
      additionalProperties: false
    ListAuditEntriesResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditEntry'
        nextCursor:
          nullable: true
          description: null 表示已到末页
          type: string
          minLength: 1
          maxLength: 512
      required:
        - items
        - nextCursor
      additionalProperties: false
...
    ListRolesResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Role'
      required:
        - items
      additionalProperties: false
    UpdateRolePermissionsRequest:
      type: object
      properties:
        rowRule:
          allOf:
            - $ref: '#/components/schemas/RowRule'
        visibleFields:
          type: array
          items:
            type: string
            enum:
              - cost
              - margin
              - price
              - subject
              - staff
        allowedActions:
          type: array
          items:
            $ref: '#/components/schemas/ActionKey'
        modules:
          type: array
          items:
            type: string
        reason:
          type: string
          minLength: 4
          maxLength: 500
          description: 变更原因。会写入审计轨迹；核查时真正被问的就是这一栏。
      required:
        - reason
      additionalProperties: false
...
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://sso.example/.well-known/openid-configuration
      description: 内部员工：企业微信 / 飞书 OIDC
    magicLink:
      type: http
      scheme: bearer
      description: 外部方（机构办 / 研究者）：一次性魔法链接换取的短期会话令牌
security:
  - oidc: []
  - magicLink: []